+++ app/Http/Controllers/TagController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Http;
+use App\Http\Requests\vertifyRequest;
 
-class TagsController extends Controller
+class TagController extends Controller
 {
     private  string $API_URL;
 
@@ -22,7 +22,7 @@ class TagsController extends Controller
         return response()->json($tags);
     }
 
-    public function store(Request $request)
+    public function store(vertifyRequest $request)
     {
         $data = $request->all();
         $response = Http::post($this->API_URL, $data);
@@ -36,17 +36,19 @@ class TagsController extends Controller
         return response()->json($tags);
     }
 
-    public function update(Request $request, string $id)
+    public function update(vertifyRequest $request, string $id)
     {
         $response = Http::patch("{$this->API_URL}/{$id}", $request);
         $tags = $response->json();
         return response()->json($tags);
     }
 
-    public function destroy(string $id)
+    public function destroy(vertifyRequest $request, string $id)
     {
+        if($request){
         $response = Http::delete("{$this->API_URL}/{$id}");
         $tags = $response->json();
-        return response()->json($tags);
+        return response()->json($tags);  
+        }
     }
 }
